pyproject.toml
[tool.poetry]
name = "la-auction"
version = "0.1.0"
description = "Verify Documents and Store"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

packages = [
    { include = "app" }
]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
tortoise-orm = "^0.25.0"
pydantic = {extras = ["dotenv", "email"], version = "^2.11.3"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
aiokafka = "^0.12.0"
python-multipart = "^0.0.20"
uvicorn = "^0.34.1"
asyncpg = "^0.30.0"
celery = {extras = ["redis"], version = "^5.5.1"}
boto3 = "^1.37.34"
typing-extensions = "^4.13.2"
phonenumbers = "^9.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.3.0"
loguru = "^0.7.3"
pydantic-settings = "^2.8.1"
confluent-kafka = "^2.9.0"
pillow = "^11.2.1"
clamd = "^1.0.2"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
